+++ pkgs/code_assets/example/stb_image/tool/ffigen.dart
@@ -6,9 +6,9 @@ import 'dart:io';
 
 import 'package:ffigen/ffigen.dart';
 
-void main() {
+Future<void> main() async {
   final packageRoot = Platform.script.resolve('../');
-  FfiGenerator(
+  final generator = FfiGenerator(
     input: Input(entryPoints: [packageRoot.resolve('third_party/stb_image.h')]),
     visitors: [
       Visitor(
@@ -19,7 +19,9 @@ void main() {
       ),
     ],
     output: Output(
-      dartFile: packageRoot.resolve('lib/src/third_party/stb_image.g.dart'),
+      dart: DartOutput(
+        path: packageRoot.resolve('lib/src/third_party/stb_image.g.dart'),
+      ),
       recordUseMapping: packageRoot.resolve(
         'lib/src/third_party/record_use_mapping.dart',
       ),
@@ -42,5 +44,6 @@ void main() {
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ''',
     ),
-  ).generate();
+  );
+  await generator.generate();
 }