+++ pkgs/ffigen/test/native_objc_test/swift_unavailable_test.dart
@@ -18,15 +18,17 @@ void main() {
   group('swift_unavailable', () {
     late final String bindings;
 
-    setUpAll(() {
-      FfiGenerator(
+    setUpAll(() async {
+      final generator = FfiGenerator(
         output: Output(
-          dartFile: Uri.file(
-            path.join(
-              packagePathForTests,
-              'test',
-              'native_objc_test',
-              'swift_unavailable_test_bindings.dart',
+          dart: DartOutput(
+            path: Uri.file(
+              path.join(
+                packagePathForTests,
+                'test',
+                'native_objc_test',
+                'swift_unavailable_test_bindings.dart',
+              ),
             ),
           ),
           format: false,
@@ -54,7 +56,8 @@ void main() {
                 node.isIncluded = node.originalName == 'Animal',
           ),
         ],
-      ).generate(logger: createTestLogger());
+      );
+      await generator.generate(logger: createTestLogger());
       final file = path.join(
         packagePathForTests,
         'test',