+++ pkgs/jni_flutter/tool/generate_jni_bindings.dart
@@ -24,7 +24,7 @@ Future<void> main() async {
       ),
     ),
     output: Output(
-      dart: DartCodeOutput(
+      dart: DartOutput(
         path: packageRoot.resolve('lib/src/generated_plugin.dart'),
         structure: OutputStructure.singleFile,
       ),